Move builder/builder-next to daemon/internal/builder-next
