+++ daemon/command/daemon.go
@@ -19,8 +19,6 @@ import (
 	"github.com/containerd/containerd/v2/pkg/tracing"
 	"github.com/containerd/log"
 	"github.com/docker/docker/api"
-	buildkit "github.com/docker/docker/builder/builder-next"
-	"github.com/docker/docker/builder/builder-next/exporter"
 	"github.com/docker/docker/builder/dockerfile"
 	"github.com/docker/docker/daemon"
 	buildbackend "github.com/docker/docker/daemon/build"
@@ -28,6 +26,8 @@ import (
 	"github.com/docker/docker/daemon/command/debug"
 	"github.com/docker/docker/daemon/command/trap"
 	"github.com/docker/docker/daemon/config"
+	buildkit "github.com/docker/docker/daemon/internal/builder-next"
+	"github.com/docker/docker/daemon/internal/builder-next/exporter"
 	"github.com/docker/docker/daemon/internal/libcontainerd/supervisor"
 	"github.com/docker/docker/daemon/listeners"
 	dopts "github.com/docker/docker/daemon/pkg/opts"
