+++ daemon/internal/builder-next/adapters/containerimage/pull.go
