+++ daemon/internal/builder-next/adapters/snapshot/layer.go
