+++ daemon/internal/builder-next/adapters/snapshot/snapshot.go
