+++ daemon/internal/builder-next/builder.go
@@ -18,11 +18,11 @@ import (
 	"github.com/docker/docker/api/types/network"
 	timetypes "github.com/docker/docker/api/types/time"
 	"github.com/docker/docker/builder"
-	"github.com/docker/docker/builder/builder-next/exporter"
-	"github.com/docker/docker/builder/builder-next/exporter/mobyexporter"
-	"github.com/docker/docker/builder/builder-next/exporter/overrides"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/images"
+	"github.com/docker/docker/daemon/internal/builder-next/exporter"
+	"github.com/docker/docker/daemon/internal/builder-next/exporter/mobyexporter"
+	"github.com/docker/docker/daemon/internal/builder-next/exporter/overrides"
 	"github.com/docker/docker/daemon/libnetwork"
 	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/errdefs"
