+++ daemon/internal/builder-next/controller.go
@@ -18,15 +18,15 @@ import (
 	"github.com/containerd/platforms"
 	"github.com/docker/docker/api/types/build"
 	"github.com/docker/docker/api/types/filters"
-	"github.com/docker/docker/builder/builder-next/adapters/containerimage"
-	"github.com/docker/docker/builder/builder-next/adapters/localinlinecache"
-	"github.com/docker/docker/builder/builder-next/adapters/snapshot"
-	"github.com/docker/docker/builder/builder-next/exporter/mobyexporter"
-	"github.com/docker/docker/builder/builder-next/imagerefchecker"
-	mobyworker "github.com/docker/docker/builder/builder-next/worker"
-	wlabel "github.com/docker/docker/builder/builder-next/worker/label"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/graphdriver"
+	"github.com/docker/docker/daemon/internal/builder-next/adapters/containerimage"
+	"github.com/docker/docker/daemon/internal/builder-next/adapters/localinlinecache"
+	"github.com/docker/docker/daemon/internal/builder-next/adapters/snapshot"
+	"github.com/docker/docker/daemon/internal/builder-next/exporter/mobyexporter"
+	"github.com/docker/docker/daemon/internal/builder-next/imagerefchecker"
+	mobyworker "github.com/docker/docker/daemon/internal/builder-next/worker"
+	wlabel "github.com/docker/docker/daemon/internal/builder-next/worker/label"
 	"github.com/docker/go-units"
 	"github.com/moby/buildkit/cache"
 	"github.com/moby/buildkit/cache/metadata"
