+++ daemon/internal/builder-next/executor_windows.go
