+++ daemon/internal/builder-next/exporter/mobyexporter/writer.go
