+++ daemon/internal/builder-next/imagerefchecker/checker.go
