+++ daemon/internal/builder-next/worker/worker.go
@@ -14,8 +14,8 @@ import (
 	cerrdefs "github.com/containerd/errdefs"
 	"github.com/containerd/log"
 	"github.com/containerd/platforms"
-	imageadapter "github.com/docker/docker/builder/builder-next/adapters/containerimage"
-	mobyexporter "github.com/docker/docker/builder/builder-next/exporter"
+	imageadapter "github.com/docker/docker/daemon/internal/builder-next/adapters/containerimage"
+	mobyexporter "github.com/docker/docker/daemon/internal/builder-next/exporter"
 	distmetadata "github.com/docker/docker/distribution/metadata"
 	"github.com/docker/docker/distribution/xfer"
 	"github.com/docker/docker/internal/mod"
